pyproject.toml
[tool.poetry]
name = "foxvoid"
version = "0.1.0"
description = ""
authors = ["ikigami0513 <ikigami0513@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.8"
django-unfold = "^0.71.0"
pillow = "^12.0.0"
python-dotenv = "^1.2.1"
mysqlclient = "^2.2.7"
django-modeltranslation = "^0.19.17"
django-countries = "^8.1.1"
django-timezone-field = "^7.1"
django-debug-toolbar = "^6.1.0"
django-tailwind = {extras = ["reload"], version = "^4.4.1"}
django-rosetta = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
